Week 8 gets data from alpha vantage, and gets either the latest stock quote from its json, or a list of stock quotes from its json.  Sort of.
It's not what I would call "production ready", because I don't utilize the interval, or the datFrom and dateTo params,
but that's just more java, and not the concept of getting data from a rest call.

Also, the parsed json isn't sorted.  So to really get the latest quote, I'd have to sort it, descending.

Also, I didn't use a serializable object.

So, some effort, some code, and I used a real rest call, which was cool, because we need to start doing that at work
-- that is, writing restful web apps or whatever.